+++ module-order-service/src/main/java/com/fashionmall/order/dto/response/common/IamPortResponseDto.java
@@ -1,4 +1,4 @@
-package com.fashionmall.common.response;
+package com.fashionmall.order.dto.response.common;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
